requirements.txt
NumPy
tgcrypto
Pillow
pyrogram
glitch_this
requests
opencv-python-headless
googletrans==3.1.0a0
